Change e-mail-address to email-address in CSS
There was a problem where an administrator could not click on the
username field to enter a username on the manage user screen. This was
caused by the email field's class being `email-address-field` while the CSS
was referring to `e-mail-address-field`. Since the style rules were not
applied correctly, the email field "overlapped" the username field. This
changes the CSS rules to match the correct `email-address-field` class.